refactor: Add utils/query
