+++ src/middleware/comments.ts
@@ -1,8 +1,10 @@
 import express from 'express'
 import CommentModel, { Comment } from '../db/models/comments'
 import { saveComment } from '../core/comments'
-import { setParams, paginate, setQuery } from '../utils/paginate'
-import type { ParamsPayload, Paginate, CriteriaPayload } from '../utils/paginate'
+import { paginate } from '../utils/paginate'
+import type { Paginate } from '../utils/paginate'
+import { applyParamsToQuery, setParams } from '../utils/query'
+import type { ParamsPayload } from '../utils/query'
 
 type CreatePayload = {
   comment: string
@@ -11,7 +13,7 @@ type CreatePayload = {
 }
 
 const index = async (
-  req: express.Request<{ projectId: string }, {}, {}, CriteriaPayload>,
+  req: express.Request<{ projectId: string }, {}, {}, ParamsPayload>,
   res: express.Response<Paginate<Comment>>,
   next: express.NextFunction
 ) => {
@@ -27,7 +29,7 @@ const index = async (
 
     const [total, data] = await Promise.all([
       CommentModel.countDocuments(criteria),
-      setQuery(CommentModel.find(criteria), params),
+      applyParamsToQuery(CommentModel.find(criteria), params),
     ])
 
     res.status(200).json(paginate({ ...params, total }, data))
@@ -45,7 +47,7 @@ const create = async (
     const comment = new CommentModel({ ...req.body, project: req._project._id })
     await saveComment(comment, req._user)
 
-    index(<express.Request<{ projectId: string }>>req, res, next)
+    index(<express.Request<{ projectId: string }, {}, {}, ParamsPayload>>req, res, next)
   } catch (error) {
     next(error)
   }
